apps/requirements.txt
Django==4.2
djangorestframework
drf-yasg
celery==5.3
redis
moviepy
django-celery-beat
flower
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-django
opentelemetry-instrumentation-requests
psycopg[binary]>=3.1
pillow<10.0
imageio-ffmpeg==0.2.0
opentelemetry-instrumentation-celery



